requirements.txt
arrow==1.2.3
binaryornot==0.4.4
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cookiecutter==2.1.1
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.2
Mesa==1.1.1
networkx==3.0
numpy==1.24.2
pandas==1.5.3
python-dateutil==2.8.2
python-slugify==8.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
text-unidecode==1.3
tornado==6.2
tqdm==4.64.1
urllib3==1.26.14
